move intergreen classes into correct package
