+++ contribs/signals/src/main/java/org/matsim/contrib/signals/data/SignalsData.java
@@ -20,7 +20,7 @@
 package org.matsim.contrib.signals.data;
 
 import org.matsim.contrib.signals.data.ambertimes.v10.AmberTimesData;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreenTimesData;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesData;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalControlData;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalGroupsData;
 import org.matsim.contrib.signals.data.signalsystems.v20.SignalSystemsData;
