+++ contribs/signals/src/main/java/org/matsim/contrib/signals/data/SignalsScenarioWriter.java
@@ -22,13 +22,13 @@ package org.matsim.contrib.signals.data;
 import org.apache.log4j.Logger;
 import org.matsim.core.controler.OutputDirectoryHierarchy;
 import org.matsim.contrib.signals.data.ambertimes.v10.AmberTimesWriter10;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesData;
 import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesWriter10;
 import org.matsim.contrib.signals.data.signalcontrol.v20.SignalControlWriter20;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalGroupsWriter20;
 import org.matsim.contrib.signals.data.signalsystems.v20.SignalSystemsWriter20;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.contrib.signals.data.ambertimes.v10.AmberTimesData;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreenTimesData;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalControlData;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalGroupsData;
 import org.matsim.contrib.signals.data.signalsystems.v20.SignalSystemsData;
