+++ contribs/signals/src/main/java/org/matsim/contrib/signals/data/intergreens/v10/IntergreenTimesData.java
@@ -1,10 +1,10 @@
 /* *********************************************************************** *
  * project: org.matsim.*
- * IntergreenTimesDataFactory
+ * IntergreenTimes
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2012 by the members listed in the COPYING,        *
+ * copyright       : (C) 2010 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -17,19 +17,29 @@
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
  * *********************************************************************** */
-package org.matsim.contrib.signals.data.ambertimes.v10;
+package org.matsim.contrib.signals.data.intergreens.v10;
+
+import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.core.api.internal.MatsimFactory;
+import org.matsim.core.api.internal.MatsimToplevelContainer;
 import org.matsim.contrib.signals.model.SignalSystem;
+
 
 
 /**
  * @author dgrether
  *
  */
-public interface IntergreenTimesDataFactory extends MatsimFactory {
-
-	public IntergreensForSignalSystemData createIntergreensForSignalSystem(Id<SignalSystem> signalSystemId);
+public interface IntergreenTimesData extends MatsimToplevelContainer  {
+	
+	public IntergreensForSignalSystemData addIntergreensForSignalSystem(IntergreensForSignalSystemData intergreens);
+	
+	public Map<Id<SignalSystem>, IntergreensForSignalSystemData> getIntergreensForSignalSystemDataMap();
+	
+	@Override
+	public IntergreenTimesDataFactory getFactory() ;
+	
+	public void setFactory(IntergreenTimesDataFactory factory);
 	
 }
