+++ contribs/signals/src/main/java/org/matsim/contrib/signals/data/intergreens/v10/IntergreenTimesDataFactory.java
@@ -1,6 +1,6 @@
 /* *********************************************************************** *
  * project: org.matsim.*
- * IntergreensForSignalSystemDataImpl
+ * IntergreenTimesDataFactory
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
@@ -17,14 +17,10 @@
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
  * *********************************************************************** */
-package org.matsim.contrib.signals.data.ambertimes.v10;
-
-import java.util.List;
-import java.util.Map;
+package org.matsim.contrib.signals.data.intergreens.v10;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.core.utils.collections.Tuple;
-import org.matsim.contrib.signals.model.SignalGroup;
+import org.matsim.core.api.internal.MatsimFactory;
 import org.matsim.contrib.signals.model.SignalSystem;
 
 
@@ -32,16 +28,8 @@ import org.matsim.contrib.signals.model.SignalSystem;
  * @author dgrether
  *
  */
-public interface IntergreensForSignalSystemData {
+public interface IntergreenTimesDataFactory extends MatsimFactory {
 
-	public Id<SignalSystem> getSignalSystemId();
-	
-	public Integer getIntergreenTime(Id<SignalGroup> endingSignalGroupId, Id<SignalGroup> beginningSignalGroupId);
-
-	public void setIntergreenTime(Integer timeSeconds, Id<SignalGroup> endingSignalGroupId, Id<SignalGroup> beginningSignalGroupId);
-
-	public List<Tuple<Id<SignalGroup>, Id<SignalGroup>>> getEndingBeginningSignalGroupKeys();
-
-	public Map<Id<SignalGroup>, Integer> getEndSignalGroupTimesForBeginningGroup(Id<SignalGroup> id);
+	public IntergreensForSignalSystemData createIntergreensForSignalSystem(Id<SignalSystem> signalSystemId);
 	
 }
