+++ contribs/signals/src/main/java/org/matsim/contrib/signals/data/intergreens/v10/IntergreensForSignalSystemDataImpl.java
@@ -27,7 +27,6 @@ import java.util.Map.Entry;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.core.utils.collections.Tuple;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreensForSignalSystemData;
 import org.matsim.contrib.signals.model.SignalGroup;
 import org.matsim.contrib.signals.model.SignalSystem;
