+++ contribs/signals/src/main/java/org/matsim/contrib/signals/model/IntergreensLogicImpl.java
@@ -28,8 +28,8 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.signals.events.SignalGroupStateChangedEvent;
 import org.matsim.core.mobsim.qsim.interfaces.SignalGroupState;
 import org.matsim.contrib.signals.SignalSystemsConfigGroup;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreenTimesData;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreensForSignalSystemData;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesData;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreensForSignalSystemData;
 
 
 /**
