+++ contribs/signals/src/main/java/tutorial/fixedTimeSignals/CreateIntergreensExample.java
@@ -24,13 +24,13 @@ import java.io.IOException;
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.signals.data.SignalsDataLoader;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesData;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesDataFactory;
 import org.matsim.contrib.signals.data.intergreens.v10.IntergreenTimesWriter10;
+import org.matsim.contrib.signals.data.intergreens.v10.IntergreensForSignalSystemData;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.contrib.signals.data.SignalsData;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreenTimesData;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreenTimesDataFactory;
-import org.matsim.contrib.signals.data.ambertimes.v10.IntergreensForSignalSystemData;
 import org.matsim.contrib.signals.model.SignalGroup;
 import org.matsim.contrib.signals.model.SignalSystem;
 import org.matsim.contrib.signals.SignalSystemsConfigGroup;
